requirements.txt
beautifulsoup4==4.7.1
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.7
django-bootstrap4==0.0.7
django-filter==2.1.0
django-friendship==1.8.0
django-heroku==0.3.1
django-widget-tweaks==1.4.3
future==0.16.0
gunicorn==19.9.0
idna==2.8
numpy==1.15.2
oauthlib==3.0.1
Pillow==5.4.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
pygame==1.9.3
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.0.0
soupsieve==1.8
urllib3==1.24.1
whitenoise==4.1.2
